intial commit for resoure gather
All of the code is in place to allow resouce gathering to work, I think.
I was relying on engine.map.tiles[x,y].tile_type to be fill but currently
it is not being full in any way so my code never executes.

If .tile_type gets filled my code should start working. If we decide
that we want to pull the info from somewhere else I will have to redo
some of my code.